+++ src/client/js/stores/texteditor.js
@@ -9,26 +9,7 @@ import ImageEntity from '../EditorEntities/ImageEntitiy.js';
 import SocialMediaEntity from '../EditorEntities/SocialMediaEntity.js';
 import JSONDocument from '../JSONDocument.js';
 import WrioActions from '../actions/wrio.js';
-import {getImageObject} from '../JSONDocument.js';
-
-// helper function
-const findEntitiesOfType = (type) => (contentBlock, callback) => {
-    contentBlock.findEntityRanges(
-        (character) => {
-            const entityKey = character.getEntity();
-            return (
-                !!entityKey &&
-                Entity.get(entityKey).getType() === type
-            );
-        },
-        callback
-    );
-};
-
-const findLinkEntities   = findEntitiesOfType('LINK');
-const findImageEntities  = findEntitiesOfType('IMAGE');
-const findSocialEntities = findEntitiesOfType('SOCIAL');
-const isImageLink = (filename) => (/\.(gif|jpg|jpeg|tiff|png)$/i).test(filename);
+import EntityTools,{getSelection,findLinkEntities,findImageEntities,findSocialEntities} from '../utils/entitytools.js';
 
 
 function appendHttp(url) {
@@ -50,36 +31,15 @@ export default Reflux.createStore({
     },
 
     setLinkEditCallback(cb) {
-        this.state.linkEditCallback = cb;
+        EntityTools.setLinkEditCallback(cb);
     },
     setImageEditCallback(cb) {
-        this.state.imageEditCallback = cb;
-    },
-
-    createLinkEntity(title,url,desc) {
-        return Entity.create('LINK', 'MUTABLE', {
-            linkTitle: title,
-            linkUrl: url,
-            linkDesc: desc,
-            editCallback: this.state.linkEditCallback
-        });
+        EntityTools.setImageEditCallback(cb);
     },
 
     getSelectedText() {
         const { editorState } = this.state;
-        var title = '';
-        const selectionState = editorState.getSelection();
-        const blockKey = selectionState.getAnchorKey();
-        const contentBlocks = editorState.getCurrentContent().getBlocksAsArray();
-        var start = selectionState.getStartOffset();
-        var end = selectionState.getEndOffset();
-
-        contentBlocks.forEach((block) => {
-            if(block.key === blockKey){
-                title = block.text.slice(start, end);
-            }
-        });
-        return title;
+        return getSelection(editorState);
     },
 
     onUpdateEditorState(state) {
@@ -94,7 +54,6 @@ export default Reflux.createStore({
     onPublishEditorState(state) {
         this.state.editorState = state;
         this.trigger(this.state);
-      //  console.log("reaction",state);
     },
 
     createEditorState(metaBlocks, mentions, images) {
@@ -125,78 +84,20 @@ export default Reflux.createStore({
             EditorState.createWithContent(ContentState.createFromBlockArray(contentBlocks), decorator) :
             EditorState.createEmpty(decorator);
 
-        editorState = metaBlocks.reduce((editorState,metaBlock) => metaBlock.data ? this.constructSocial(editorState,metaBlock) : editorState, editorState);
+        editorState = metaBlocks.reduce((editorState,metaBlock) => metaBlock.data ? EntityTools.constructSocial(editorState,metaBlock) : editorState, editorState);
         if (images) {
-            editorState = images.reduce((editorState,mention) => this.constructImage(editorState,orderedBlocks,mention),editorState);
+            editorState = images.reduce((editorState,mention) => EntityTools.constructImage(editorState,orderedBlocks,mention),editorState);
         }
 
-        return mentions.reduce((editorState,mention) => this.constructMention(editorState,orderedBlocks,mention),editorState);
+        return mentions.reduce((editorState,mention) => EntityTools.constructMention(editorState,orderedBlocks,mention),editorState);
 
     },
-
-    constructSocial(editorState,metaBlock) {
-        const contentBlock = metaBlock.block;
-        const blockData = metaBlock.data;
-        let entityKey;
-        if (blockData["@type"] == "ImageObject") {
-            entityKey = this._createImageSocialEntity(blockData.contentUrl,blockData.name,blockData.description);
-        } else {
-            entityKey = this._createImageSocialEntity(blockData.sharedContent.url,blockData.sharedContent.headline,block.sharedContent.about);
-        }
-        const key = contentBlock.getKey();
-        const _editorState = EditorState.forceSelection(editorState,SelectionState.createEmpty(key));
-        return this._insertEntityKey(_editorState,entityKey);
-
-    },
-
-    _getMentionContentBlock(contentBlocks,mention) {
-        const block = contentBlocks[mention.block];
-
-        if (!block) {
-            console.warn("Cannot create mention",mention);
-            throw new Error("Mention create error");
-        }
-        return block;
-    },
-
-    constructEntity(entityKey,editorState,contentBlocks,mention) {
-
-        try {
-            const key = this._getMentionContentBlock(contentBlocks,mention).getKey();
-            return RichUtils.toggleLink(
-                editorState,
-                SelectionState.createEmpty(key).merge({
-                    anchorOffset: mention.start,
-                    focusKey: key,
-                    focusOffset: mention.end
-                }),
-                entityKey
-            );
-        } catch (e){
-            console.error("Error mapping a mention",e);
-            return editorState;
-        }
-    },
-
-    constructMention(editorState, contentBlocks, mention) {
-        const entityKey = this.createLinkEntity(mention.linkWord,mention.url,mention.linkDesc);
-        return this.constructEntity(entityKey,editorState,contentBlocks,mention);
-    },
-
-    constructImage(editorState, contentBlocks, mention) {
-       const metaData = {
-           block: this._getMentionContentBlock(contentBlocks,mention),
-           data: getImageObject(mention.src,mention.name,mention.description)
-       };
-       return this.constructSocial(editorState,metaData);
-    },
-
 
     onCreateNewLink(titleValue,urlValue,descValue) {
 
         urlValue = appendHttp(urlValue);
 
-        const entityKey = this.createLinkEntity(titleValue,urlValue,descValue);
+        const entityKey = EntityTools.createLinkEntity(titleValue,urlValue,descValue);
         const {editorState} = this.state;
 
         const e = Entity.get(entityKey).getData();
@@ -210,6 +111,12 @@ export default Reflux.createStore({
         this.onPublishEditorState(_editorState);
     },
 
+    onCreateNewImage (url,description,title) {
+        const entityKey = EntityTools.createImageSocialEntity(url,description,title);
+        const {editorState} = this.state;
+        this.onPublishEditorState(EntityTools.insertEntityKey(editorState,entityKey));
+    },
+
     onEditLink(titleValue,urlValue,descValue,linkEntityKey) {
         Entity.mergeData(linkEntityKey, {
             linkTitle: titleValue,
@@ -218,44 +125,13 @@ export default Reflux.createStore({
         });
     },
 
-    onCreateNewImage (url,description,title) {
-        const entityKey = this._createImageSocialEntity(url,description,title);
-        const {editorState} = this.state;
-        this.onPublishEditorState(this._insertEntityKey(editorState,entityKey));
-    },
-
-    _createImageSocialEntity(url,description,title) {
-        const urlType = isImageLink(url) ? 'IMAGE' : 'SOCIAL';
-        const entityKey = Entity.create(urlType, 'IMMUTABLE',
-            {
-                src: url ,
-                description,
-                title,
-                editCallback: this.state.imageEditCallback
-            });
-        return entityKey;
-    },
-
-    _insertEntityKey(editorState, entityKey) {
-        const newEditorState = AtomicBlockUtils.insertAtomicBlock(
-            editorState,
-            entityKey,
-            ' '
-        );
-        return EditorState.forceSelection(
-            newEditorState,
-            editorState.getCurrentContent().getSelectionAfter()
-        );
-    },
-
     onEditImage(src,description,title,linkEntityKey) {
         Entity.mergeData(linkEntityKey, {
             src,
+            title,
             description
         });
     },
-
-
 
     onRemoveLink(linkEntityKey) {
         const {editorState} = this.state;
